Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a,b;
-    scanf("%d%d",&a,&b);
-    printf("%d%d",a*b);
+    float a,b;
+    scanf("%f%f",&a,&b);
+    printf("%.2f%.2f",a*b);
     
     return 0;
 }
